Refactored Project
