--- frontend/src/App.css
@@ -1,102 +1,0 @@
-.center {
-  width: 50%;
-  margin: auto;
-  color: white;
-  font-family: Arial, serif;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 1700ms linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.deleteBtn {
-  cursor: pointer;
-  border: 2px solid red;
-  background-color: transparent;
-  color: red;
-  font-size: 25px;
-  padding: 8px;
-}
-
-.label, .input {
-  text-align: left;
-  padding: 8px;
-  font-family: "Comic Sans MS", cursive;
-;
-  font-size: 35px;
-}
-
-input[type=checkbox]:checked + label {
-  text-decoration: line-through;
-  color: #8d8d8d;
-}
-
-.table {
-  padding: 5px;
-  margin: 1px;
-}
-
-.headline {
-  color: white;
-  font-size: 40px;
-  font-family: Intim, sans-serif;
-}
-
-.textbar {
-  padding: 5px;
-  font-family: "Comic Sans MS", cursive;
-  font-size: 35px;
-  color: white;
-  background-color: #282c34;
-  border: 0;
-  border-bottom: 2px solid white;
-}
-
-.btn {
-  padding: 8px;
-  font-family: Intim, sans-serif;
-  cursor: pointer;
-  border: 2px solid deepskyblue;
-  background-color: transparent;
-  color: deepskyblue;
-  font-size: 25px;
-  margin-left: 10px;
-  width: 150px;
-}
-
-::selection {
-  background: deepskyblue;
-}
-
-.img {
-  max-width: 20%;
-  max-height: 20%;
-}
-
-.div1 {
-  display: flex;
-  justify-content: space-between;
-}
