--- frontend/src/index.css
@@ -1,15 +1,0 @@
-
-.header {
-  font-size: 80px;
-  color: white;
-  font-family: "Inter Semi Bold", sans-serif;
-}
-
-code {
-
-}
-
-.div {
-  width: 50%;
-  margin: auto;
-}
