--- package.json
@@ -1,8 +1,0 @@
-{
-  "devDependencies": {
-    "@graphql-codegen/cli": "^5.0.2",
-    "@graphql-codegen/client-preset": "^4.2.3",
-    "@graphql-typed-document-node/core": "^3.2.0",
-    "typescript": "^5.3.3"
-  }
-}
