+++ todolist-be/src/main/resources/db/migration/V003__rename-database.sql
