+++ todolist-fe/public/favicon.ico
