+++ todolist-fe/src/TodoItem.tsx
@@ -4,9 +4,8 @@ import {
     useGetItemQuery,
     usePutItemMutation
 } from "./graphql/generated/graphql";
-import './App.css'
+import './styles/App.css'
 import React, {useEffect, useState} from "react";
-import {TodoItem} from "./App";
 
 const GetTodoItem = () => {
 
@@ -18,40 +17,48 @@ const GetTodoItem = () => {
     const {data, loading, refetch} = useGetItemQuery()
 
     useEffect(() => {
-        if (loading) {
-            console.log(loading)
-        } else {
-            setItems(data?.getItem)
-        }
-    }, [data, loading, items]);
+        loading ? console.log(loading) : setItems(data?.getItem)
+    }, [data, items]);
 
     function changeIsDone(isDone: boolean, uuid: string) {
         isDone = !isDone
-        putItem({variables: {uuid: uuid, isDone: isDone}}).then(r =>
+        putItem({variables: {uuid: uuid, isDone: isDone}}).then(() =>
             refetch()
-        )
+        ).catch(e => {
+            alert(e)
+        })
     }
 
     function deleteItem(uuid: string) {
-        removeItem({variables: {uuid: uuid}}).then(r =>
+        removeItem({variables: {uuid: uuid}}).then(() =>
             refetch()
-        )
+        ).catch(e => {
+            alert(e)
+        })
     }
 
     function createTodoItem() {
-        createItem({variables: {item: item}}).then(r =>
-            refetch()
-        ).catch()
+        if (item === "") {
+            console.log(item)
+        } else {
+            createItem({variables: {item: item}}).then(() =>
+                refetch()
+            ).catch(e => {
+                alert(e)
+            })
+        }
     }
 
     return (
         <>{
-            loading ? <div className="center"> <img src="https://static-00.iconduck.com/assets.00/spinner-icon-2048x2048-mhj15xft.png" className="App-logo img"/> </div> :
+            loading ? <div className="center"><img alt="loading"
+                    src="https://static-00.iconduck.com/assets.00/spinner-icon-2048x2048-mhj15xft.png"
+                    className="App-logo img"/></div> :
                 <div className="center">
                     {items.map((item: any, index: number) => {
                         return (
                             <div key={index} className="table">
-                                <div className="div1">
+                                <div className="flex-between">
                                     <div>
                                         <input type="checkbox" className="input"
                                                onClick={() => changeIsDone(item?.isDone, item.uuid)}
@@ -60,24 +67,27 @@ const GetTodoItem = () => {
                                     </div>
                                     <input onClick={() => deleteItem(item?.uuid)} type="button" className="deleteBtn"
                                            value="Löschen"></input>
-
                                 </div>
-
                                 <hr/>
                             </div>
                         )
                     })}
                     <h1 className="headline">Neuer Artikel</h1>
-                    <input onKeyDown={event => {
-                        if (event.key == 'Enter') {
-                           createTodoItem()
-                        }
-                    }}
-                           onChange={event => {
-                               setItem(event.target.value)
-                           }}
-                           id="demo" className="textbar" type="text" placeholder="Neuer Artikel eingeben" value={item}/>
-                    <input type="button" onClick={createTodoItem} className="btn" value="Hinzufügen"></input>
+                        <div className="flex-between">
+                            <input onKeyDown={event => {
+                                if (event.key == 'Enter') {
+                                    createTodoItem()
+                                    setItem("")
+                                }
+                            }}
+                                   onChange={event => {
+                                       setItem(event.target.value)
+                                   }}
+                                   id="demo" className="textbar" type="text" placeholder="Neuer Artikel eingeben"
+                                   value={item}/>
+
+                            <input type="button" onClick={createTodoItem} className="btn" value="Hinzufügen"></input>
+                        </div>
                 </div>
         }
         </>
