+++ todolist-fe/src/styles/App.css
@@ -1,0 +1,121 @@
+@media (min-width: 700px) {
+    .center {
+        width: 50%;
+        margin: auto;
+        color: white;
+        font-family: Arial, serif;
+    }
+}
+@media (max-width: 1000px) {
+    .center {
+        color: white;
+        font-family: Arial, serif;
+    }
+}
+
+
+@media (prefers-reduced-motion: no-preference) {
+    .App-logo {
+        animation: App-logo-spin infinite 1700ms linear;
+    }
+}
+
+.App-header {
+    background-color: #282c34;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    font-size: calc(10px + 2vmin);
+    color: white;
+}
+
+@keyframes App-logo-spin {
+    from {
+        transform: rotate(0deg);
+    }
+
+    to {
+        transform: rotate(360deg);
+    }
+}
+
+.deleteBtn {
+    cursor: pointer;
+    border: 1px solid red;
+    background-color: transparent;
+    color: red;
+    font-size: 16px;
+    padding: 3px;
+}
+
+.label, .input {
+    text-align: left;
+    padding: 6px;
+    font-family: "Comic Sans MS", cursive;;
+    font-size: 15px;
+}
+
+input[type=checkbox]:checked + label {
+    text-decoration: line-through;
+    color: #8d8d8d;
+}
+
+input[type=checkbox] {
+    width: 8px;
+    height: 8px;
+    margin: 1px;
+}
+
+.table {
+    padding: 5px;
+    margin: 1px;
+}
+
+.headline {
+    color: white;
+    font-size: 20px;
+    font-family: Intim, sans-serif;
+}
+
+.textbar {
+    padding: 5px;
+    font-family: "Comic Sans MS", cursive;
+    font-size: 16px;
+    color: white;
+    background-color: #282c34;
+    border: 0;
+    border-bottom: 1px solid white;
+    width: 100%;
+}
+
+.btn {
+    padding: 3px;
+    font-family: Intim, sans-serif;
+    cursor: pointer;
+    border: 1px solid deepskyblue;
+    background-color: transparent;
+    color: deepskyblue;
+    font-size: 16px;
+    margin-left: 5px;
+    width: 13%;
+}
+
+::selection {
+    background: blue;
+}
+
+.img {
+    max-width: 20%;
+    max-height: 20%;
+}
+
+.flex-between {
+    display: flex;
+    justify-content: space-between;
+}
+
+.grid {
+    display: grid;
+}
